Moved images to the public dir
