+++ src/header/Header.js
@@ -1,12 +1,10 @@
 import Navbar from "react-bootstrap/Navbar";
 import Nav from "react-bootstrap/Nav";
-import Button from "react-bootstrap/Button";
 import { NavDropdown } from "react-bootstrap";
 import Form from "react-bootstrap/Form";
 import { FormControl } from "react-bootstrap";
 
-import logo from "../../src/images/bcu_logo.webp";
-// import backgroundStrip from "../../src/images/BCU London Universe STRIP.jpg";
+import logo from "../../public/images/bcu_logo.webp";
 import "./Header.css";
 import { Link } from "@reach/router";
 
@@ -15,7 +13,7 @@ const Header = () => (
         <Navbar.Brand href="#">
             <img src={logo} alt="logo" className="brand-logo" />
         </Navbar.Brand>
-        {/* <img src={backgroundStrip} alt="background strip" className="backgroundStrip" /> */}
+        
         <Navbar.Toggle aria-controls="navbarScroll" />
         <Navbar.Collapse id="navbarScroll">
             <Nav
