+++ src/home/Home.css
@@ -29,12 +29,12 @@
     outline: none !important;
     border: var(--gold-hue);
     background-image: linear-gradient(to left,
-    var(--gold-primary),
-    var(--gold-light-1),
-    var(--gold-dark),
-    var(--gold-light-1),
-    var(--gold-primary)) !important;
-    
+        var(--gold-primary),
+        var(--gold-light-1),
+        var(--gold-dark),
+        var(--gold-light-1),
+        var(--gold-primary)) !important;
+        
     color: var(--custom-black) !important;
     font-size: 0.9rem;
     text-transform: capitalize;
@@ -180,6 +180,20 @@
     }
     .tv-box iframe {
         width: 900px !important;
+    }
+
+    ul::-webkit-scrollbar {
+        width: 0.8rem;
+        height: 0.8rem;
+    }
+    ul::-webkit-scrollbar-track {
+        background: #f1f1f1;
+        box-shadow: inset 0 0 5px grey; 
+        border-radius: 10px;
+    }
+    ul::-webkit-scrollbar-thumb {
+        background: var(--gold-primary);
+        border-radius: 10px;
     }
 }
