+++ src/index.css
@@ -32,6 +32,19 @@ body {
     background: var(--custom-black) !important;
 }
 
+body::-webkit-scrollbar {
+    width: 1rem;
+}
+body::-webkit-scrollbar-track {
+    background: #f1f1f1;
+    box-shadow: inset 0 0 5px grey; 
+    border-radius: 10px;
+}
+body::-webkit-scrollbar-thumb {
+    background: #888;
+    border-radius: 10px;
+}
+
 /* .image {
     position: absolute;
     z-index: -1;
